created new compensationController in the interest of separation of concerns
