+++ CodeChallenge/Controllers/CompensationController.cs
@@ -1,0 +1,61 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
+using CodeChallenge.Services;
+using CodeChallenge.Models;
+
+namespace CodeChallenge.Controllers
+{
+    [ApiController]
+    [Route("api/compensation")]
+    public class CompensationController : ControllerBase
+    {
+        private readonly ILogger _logger;
+        private readonly ICompensationService _compensationService;
+
+        public CompensationController(ILogger<CompensationController> logger, ICompensationService compensationService)
+        {
+            _logger = logger;
+            _compensationService = compensationService;
+        }
+
+        [HttpPost]
+        public IActionResult CreateEmployeeCompensation([FromBody] Compensation compensation)
+        {
+            _logger.LogDebug($"Received compensation create request for '{compensation.Employee.FirstName} {compensation.Employee.LastName}'");
+
+            _compensationService.Create(compensation);
+
+            return CreatedAtRoute("getCompensationById", new { id = compensation.CompensationId }, compensation);
+        }
+
+        [HttpGet("{id}", Name = "getCompensationById")]
+        public IActionResult GetCompensationById(String id)
+        {
+            _logger.LogDebug($"Received compensation get request for '{id}'");
+
+            var compensation = _compensationService.GetById(id);
+
+            if (compensation == null)
+                return NotFound();
+
+            return Ok(compensation);
+        }
+
+        [HttpGet("byEmployee/{employeeId}", Name = "getCompensationByEmployeeId")]
+        public IActionResult GetCompensationByEmployeeId(String employeeId)
+        {
+            _logger.LogDebug($"Received compensation get request for employee '{employeeId}'");
+
+            var compensation = _compensationService.GetByEmployeeId(employeeId);
+
+            if (compensation == null)
+                return NotFound();
+
+            return Ok(compensation);
+        }
+    }
+}
